+++ mod-source-record-manager-server/src/test/java/org/folio/services/afterprocessing/AdditionalFieldsUtilTest.java
@@ -1,10 +1,9 @@
 package org.folio.services.afterprocessing;
 
-
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
-import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.folio.TestUtil;
 import org.folio.rest.jaxrs.model.ParsedRecord;
 import org.folio.rest.jaxrs.model.Record;
 import org.junit.Assert;
@@ -12,14 +11,13 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.BlockJUnit4ClassRunner;
 
-import java.io.File;
 import java.io.IOException;
 import java.util.UUID;
 
 @RunWith(BlockJUnit4ClassRunner.class)
 public class AdditionalFieldsUtilTest {
+
   private static final String PARSED_RECORD_PATH = "src/test/resources/org/folio/services/afterprocessing/parsedRecord.json";
-  private AdditionalFieldsUtil additionalFieldsUtil = new AdditionalFieldsUtil();
 
   @Test
   public void shouldAddInstanceIdSubfield() throws IOException {
@@ -27,26 +25,35 @@ public class AdditionalFieldsUtilTest {
     String recordId = UUID.randomUUID().toString();
     String instanceId = UUID.randomUUID().toString();
 
-    String parsedRecordContent = FileUtils.readFileToString(new File(PARSED_RECORD_PATH), "UTF-8");
+    String parsedRecordContent = TestUtil.readFileFromPath(PARSED_RECORD_PATH);
     ParsedRecord parsedRecord = new ParsedRecord();
+    String leader = new JsonObject(parsedRecordContent).getString("leader");
     parsedRecord.setContent(parsedRecordContent);
     Record record = new Record().withId(recordId).withParsedRecord(parsedRecord);
     // when
-    boolean added = additionalFieldsUtil.addInstanceIdToMarcRecord(record, instanceId);
+    boolean addedSourceRecordId = AdditionalFieldsUtil.addFieldToMarcRecord(record, AdditionalFieldsUtil.TAG_999, 's', recordId);
+    boolean addedInstanceId = AdditionalFieldsUtil.addFieldToMarcRecord(record, AdditionalFieldsUtil.TAG_999, 'i', instanceId);
     // then
-    Assert.assertTrue(added);
-    JsonObject content = new JsonObject((String) record.getParsedRecord().getContent());
+    Assert.assertTrue(addedSourceRecordId);
+    Assert.assertTrue(addedInstanceId);
+    JsonObject content = new JsonObject(parsedRecord.getContent().toString());
     JsonArray fields = content.getJsonArray("fields");
+    String newLeader = content.getString("leader");
+    Assert.assertNotEquals(leader, newLeader);
     Assert.assertTrue(!fields.isEmpty());
     for (int i = fields.size(); i-- > 0; ) {
       JsonObject targetField = fields.getJsonObject(i);
-      if (targetField.containsKey(AdditionalFieldsConfig.TAG_999)) {
-        JsonArray subfields = targetField.getJsonObject(AdditionalFieldsConfig.TAG_999).getJsonArray("subfields");
+      if (targetField.containsKey(AdditionalFieldsUtil.TAG_999)) {
+        JsonArray subfields = targetField.getJsonObject(AdditionalFieldsUtil.TAG_999).getJsonArray("subfields");
         for (int j = subfields.size(); j-- > 0; ) {
           JsonObject targetSubfield = subfields.getJsonObject(j);
           if (targetSubfield.containsKey("i")) {
             String actualInstanceId = (String) targetSubfield.getValue("i");
             Assert.assertEquals(instanceId, actualInstanceId);
+          }
+          if (targetSubfield.containsKey("s")) {
+            String actualSourceRecordId = (String) targetSubfield.getValue("s");
+            Assert.assertEquals(recordId, actualSourceRecordId);
           }
         }
       }
@@ -59,7 +66,7 @@ public class AdditionalFieldsUtilTest {
     Record record = new Record();
     String instanceId = UUID.randomUUID().toString();
     // when
-    boolean added = additionalFieldsUtil.addInstanceIdToMarcRecord(record, instanceId);
+    boolean added = AdditionalFieldsUtil.addFieldToMarcRecord(record, AdditionalFieldsUtil.TAG_999, 'i', instanceId);
     // then
     Assert.assertFalse(added);
     Assert.assertNull(record.getParsedRecord());
@@ -73,7 +80,7 @@ public class AdditionalFieldsUtilTest {
     record.setParsedRecord(new ParsedRecord().withContent(content));
     String instanceId = UUID.randomUUID().toString();
     // when
-    boolean added = additionalFieldsUtil.addInstanceIdToMarcRecord(record, instanceId);
+    boolean added = AdditionalFieldsUtil.addFieldToMarcRecord(record, AdditionalFieldsUtil.TAG_999, 'i', instanceId);
     // then
     Assert.assertFalse(added);
     Assert.assertNotNull(record.getParsedRecord());
@@ -89,7 +96,7 @@ public class AdditionalFieldsUtilTest {
     record.setParsedRecord(new ParsedRecord().withContent(content));
     String instanceId = UUID.randomUUID().toString();
     // when
-    boolean added = additionalFieldsUtil.addInstanceIdToMarcRecord(record, instanceId);
+    boolean added = AdditionalFieldsUtil.addFieldToMarcRecord(record, AdditionalFieldsUtil.TAG_999, 'i', instanceId);
     // then
     Assert.assertFalse(added);
     Assert.assertNotNull(record.getParsedRecord());
@@ -105,12 +112,11 @@ public class AdditionalFieldsUtilTest {
     record.setParsedRecord(new ParsedRecord().withContent(content));
     String instanceId = UUID.randomUUID().toString();
     // when
-    boolean added = additionalFieldsUtil.addInstanceIdToMarcRecord(record, instanceId);
+    boolean added = AdditionalFieldsUtil.addFieldToMarcRecord(record, AdditionalFieldsUtil.TAG_999, 'i', instanceId);
     // then
     Assert.assertFalse(added);
     Assert.assertNotNull(record.getParsedRecord());
     Assert.assertNotNull(record.getParsedRecord().getContent());
-    Assert.assertEquals(content, record.getParsedRecord().getContent());
   }
 
   @Test
@@ -120,25 +126,10 @@ public class AdditionalFieldsUtilTest {
     record.setParsedRecord(new ParsedRecord().withContent(null));
     String instanceId = UUID.randomUUID().toString();
     // when
-    boolean added = additionalFieldsUtil.addInstanceIdToMarcRecord(record, instanceId);
+    boolean added = AdditionalFieldsUtil.addFieldToMarcRecord(record, AdditionalFieldsUtil.TAG_999, 'i', instanceId);
     // then
     Assert.assertFalse(added);
     Assert.assertNotNull(record.getParsedRecord());
     Assert.assertNull(record.getParsedRecord().getContent());
-  }
-
-  @Test
-  public void shouldNotAddInstanceIdSubfieldIfNo999Field() {
-    // given
-    Record record = new Record();
-    String content = "{\"fields\":\"[]\"}";
-    record.setParsedRecord(new ParsedRecord().withContent(content));
-    String instanceId = UUID.randomUUID().toString();
-    // when
-    boolean added = additionalFieldsUtil.addInstanceIdToMarcRecord(record, instanceId);
-    // then
-    Assert.assertFalse(added);
-    Assert.assertNotNull(record.getParsedRecord());
-    Assert.assertEquals(content, record.getParsedRecord().getContent());
   }
 }