+++ NEWS.md
@@ -8,6 +8,7 @@
  * Filtered out invalid Instances before saving to inventory
  * Added delete endpoint for job execution and all associated records from SRS
  * Added total records number to logDto
+ * Changed logic of adding fields to MARC record resulting in update of leader value
  
  | METHOD |             URL                              | DESCRIPTION                                              |
  |--------|----------------------------------------------|----------------------------------------------------------|